Let user specify tar program via TAR envar
The `--wildcards` option exists only in GNU tar. GNU tar is installed
as `gtar` on BSD systems, but as `tar` on Linux systems. Linux distros
do not typically have a `gtar` command so we can't always use `gtar`.

As a quick fix, respect the TAR environment variable which defaults to
`tar` if not given. BSD users can no do `TAR=gtar ./generate.sh`.